feat(requireUniversalModule.js): Adds option to allow applying of options in isDynamic scenarios (e.
In addModule method, we will not use the 'ignoreBabelRename' to determine if we should skip applying
a regex to the chunk name which is typically determined from the file path.  This is not always the
case, as  dynamic imports can  have  static strings,  and can also reference webpack aliases. This
gives the user more control over what they want.  Additionally,  in the getConfig method, options
provided, e.g. custom chunkName functions, would not be  applied .